requirements.txt
streamlit
pillow
numpy
moviepy
scikit-learn